ci: update teleportdao package

+++ contracts/erc20/interfaces/IWETH.sol
@@ -4,11 +4,13 @@ pragma solidity >=0.8.0 <=0.8.4;
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface IWETH is IERC20 {
-
     function name() external view returns (string memory);
+
     function symbol() external view returns (string memory);
+
     function decimals() external view returns (uint8);
 
     function deposit() external payable;
-    function withdraw(uint) external;
+
+    function withdraw(uint256) external;
 }
